# packages/api-client/openapi/dante-v1.openapi.json
{
  "components": {
    "schemas": {
      "AuthenticatedSessionResponse": {
        "additionalProperties": false,
        "description": "Authenticated bootstrap/session representation.",
        "properties": {
          "account_ref": {
            "format": "uuid",
            "title": "Account Ref",
            "type": "string"
          },
          "auth_session_ref": {
            "format": "uuid",
            "title": "Auth Session Ref",
            "type": "string"
          },
          "authenticated": {
            "const": true,
            "title": "Authenticated",
            "type": "boolean"
          },
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "recent_auth_at": {
            "format": "date-time",
            "title": "Recent Auth At",
            "type": "string"
          }
        },
        "required": [
          "account_ref",
          "auth_session_ref",
          "authenticated",
          "csrf_token",
          "expires_at",
          "recent_auth_at"
        ],
        "title": "AuthenticatedSessionResponse",
        "type": "object"
      },
      "AuthenticationMethodsResponse": {
        "additionalProperties": false,
        "description": "Current Account-wide direct-authenticator and recovery-channel inventory.",
        "properties": {
          "active_passkey_count": {
            "minimum": 0.0,
            "title": "Active Passkey Count",
            "type": "integer"
          },
          "password_established": {
            "title": "Password Established",
            "type": "boolean"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/AuthenticationProviderMethodResponse"
            },
            "title": "Providers",
            "type": "array"
          },
          "recovery_eligible_email_count": {
            "minimum": 0.0,
            "title": "Recovery Eligible Email Count",
            "type": "integer"
          }
        },
        "required": [
          "password_established",
          "providers",
          "active_passkey_count",
          "recovery_eligible_email_count"
        ],
        "title": "AuthenticationMethodsResponse",
        "type": "object"
      },
      "AuthenticationProviderMethodResponse": {
        "additionalProperties": false,
        "description": "Safe provider authenticator metadata for Security/settings surfaces.",
        "properties": {
          "external_identity_ref": {
            "format": "uuid",
            "title": "External Identity Ref",
            "type": "string"
          },
          "provider_code": {
            "title": "Provider Code",
            "type": "string"
          },
          "provider_email_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Email Address"
          },
          "provider_email_private": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Email Private"
          }
        },
        "required": [
          "external_identity_ref",
          "provider_code",
          "provider_email_address",
          "provider_email_private"
        ],
        "title": "AuthenticationProviderMethodResponse",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "description": "DANTE RFC 9457 public error contract.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemFieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "category",
          "request_id",
          "retryable"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemFieldError": {
        "additionalProperties": false,
        "description": "Stable bounded field error surfaced to first-party clients.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          },
          "pointer": {
            "title": "Pointer",
            "type": "string"
          }
        },
        "required": ["pointer", "code", "detail"],
        "title": "ProblemFieldError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "DanteSessionCookie": {
        "description": "Opaque host-only HttpOnly Secure AuthSession bearer cookie; browser transport owns it.",
        "in": "cookie",
        "name": "__Host-dante-session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "DANTE Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/methods": {
      "get": {
        "description": "Return safe Account-wide authenticator inventory for the admitted AuthSession.",
        "operationId": "auth_get_authentication_methods",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationMethodsResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "May clear an invalid or expired __Host-dante-session cookie.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "DanteSessionCookie": []
          }
        ],
        "summary": "Get Authentication Methods",
        "tags": ["auth"]
      }
    },
    "/api/v1/auth/password": {
      "delete": {
        "description": "Remove the password only when Account-wide anti-lockout remains satisfied.",
        "operationId": "auth_remove_password",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedSessionResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "Set-Cookie": {
                "description": "Rotates the bearer after password removal through __Host-dante-session.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal Server Error",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable",
            "headers": {
              "Cache-Control": {
                "description": "Auth responses are never cacheable by shared or browser caches.",
                "schema": {
                  "const": "no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "Server-authoritative non-secret request correlation identifier.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "DanteSessionCookie": []
          }
        ],
        "summary": "Remove Password",
        "tags": ["auth"],
        "x-dante-browser-security": {
          "client_header": {
            "name": "X-Dante-Client",
            "owner": "web_transport",
            "required": true,
            "required_value": "web"
          },
          "csrf_header": {
            "name": "X-Dante-CSRF",
            "owner": "web_transport",
            "required": true
          },
          "origin": {
            "must_equal": "canonical_web_origin",
            "owner": "browser",
            "required": true
          },
          "sec_fetch_site": {
            "owner": "browser",
            "required": true,
            "required_value": "same-origin"
          }
        }
      }
    },
...
  }
}