# packages/api-client/src/generated/model/authenticationMethodsResponse.zod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DANTE Backend
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod/mini';

export const authenticationMethodsResponseActivePasskeyCountMin = 0;

export const authenticationMethodsResponseRecoveryEligibleEmailCountMin = 0;

export const AuthenticationMethodsResponse = /*#__PURE__*/ zod
  .object({
    active_passkey_count: /*#__PURE__*/ zod
      .int()
      .check(
        /*#__PURE__*/ zod.gte(
          authenticationMethodsResponseActivePasskeyCountMin,
        ),
      ),
    password_established: /*#__PURE__*/ zod.boolean(),
    providers: /*#__PURE__*/ zod.array(
      /*#__PURE__*/ zod
        .object({
          external_identity_ref: /*#__PURE__*/ zod.uuid(),
          provider_code: /*#__PURE__*/ zod.string(),
          provider_email_address: /*#__PURE__*/ zod.union([
            /*#__PURE__*/ zod.string(),
            /*#__PURE__*/ zod.null(),
          ]),
          provider_email_private: /*#__PURE__*/ zod.union([
            /*#__PURE__*/ zod.boolean(),
            /*#__PURE__*/ zod.null(),
          ]),
        })
        .check(
          /*#__PURE__*/ zod.describe(
            'Safe provider authenticator metadata for Security\/settings surfaces.',
          ),
        ),
    ),
    recovery_eligible_email_count: /*#__PURE__*/ zod
      .int()
      .check(
        /*#__PURE__*/ zod.gte(
          authenticationMethodsResponseRecoveryEligibleEmailCountMin,
        ),
      ),
  })
  .check(
    /*#__PURE__*/ zod.describe(
      'Current Account-wide direct-authenticator and recovery-channel inventory.',
    ),
  );

export type AuthenticationMethodsResponse = zod.input<
  typeof AuthenticationMethodsResponse
>;
export type AuthenticationMethodsResponseOutput = zod.output<
  typeof AuthenticationMethodsResponse
>;